requirements.txt
streamlit>=1.36.0
pandas>=2.0.0
pillow>=10.0.0
pytesseract>=0.3.10
plotly>=5.20.0
openpyxl>=3.1.2
pdfplumber>=0.11.0
PyPDF2>=3.0.0
scikit-learn>=1.3.0
joblib>=1.3.2
requests>=2.31.0
